break out QuotaRequestSerializer
