+++ api/serializers/quota_request_serializer.py
@@ -1,0 +1,17 @@
+from core.models.request import QuotaRequest, StatusType
+from core.models.user import AtmosphereUser
+from rest_framework import serializers
+
+
+class QuotaRequestSerializer(serializers.ModelSerializer):
+    id = serializers.CharField(read_only=True, source="uuid", required=False)
+    created_by = serializers.SlugRelatedField(
+        slug_field='username', source='created_by',
+        queryset=AtmosphereUser.objects.all())
+    status = serializers.SlugRelatedField(
+        slug_field='name', source='status',
+        queryset=StatusType.objects.all())
+
+    class Meta:
+        model = QuotaRequest
+        exclude = ('uuid', 'membership')
